refactor component structure by adding layout, core and page folders
