+++ src/components/core/Finder/finder.test.js
